add train llm file

+++ src/train_llm.py
@@ -655,7 +655,7 @@ def main(cfg: DictConfig):
         #     bnb_4bit_compute_dtype=torch.bfloat16
         # )
 
-        base_model = LlamaForTokenClassification.from_pretrained(cfg.model_name, num_labels=len(LABELS), trust_remote_code=True, torch_dtype=torch.float16)
+        base_model = LlamaForTokenClassification.from_pretrained(cfg.model_name, num_labels=len(LABELS), trust_remote_code=True, torch_dtype=torch.bfloat16)
         base_model.config.pretraining_tp = 1
         # base_model = prepare_model_for_kbit_training(base_model)
         # for name, param in base_model.named_parameters():
